fix: rounding on interation-distribution calculations
When using  very low rates intended to only fire the scenarios
a couple times over the duration, the calculations for the
iteration distribution could fail to add up to 1 by the last
tick of the iteration.

Using `Ceil` instead of `Round` ensures 1 is always reached consistently.

See added test cases for examples that would fail to trigger the scenario
without the fix.

Co-authored-by: Kevin Intriago <kevin.intriago@form3.tech>