feat: optional id label for all metrics (#154)
Metrics are exported via prometheus push gateway. Unfortunately, push gateway is only a super simple cache. It does not add any common labels, such as server, or pod name, and when metric with the same label set is provided, it overrides the latest value with a new one. That's why it's crucial to make label values unique.

With the new id label it should be a little easier. It's optional and needed only when metric entries might be not unique.