feat: rewrite work scheduling to improve performance
Changes:
 - Separate worker pool implementation for users and iteration triggers
 - Don't use wall clock to measure iteration duration
 - Don't use channels to trigger work to avoid locking
 - remove `tracer` from the worker pool as it provides little value.
 - Add a scenario that sleeps a configurable duration to the benchmark command.

`ContinuousPool`:

The `users` trigger type executes the scenario as fasts as possible given
a number of workers. Meaning, we don't need a channel to trigger or
control the work externally. Just start a for loop to run jobs until
timeout has passed or max iterations have been reached.

`TriggerPool`:

The `iterations` trigger will try to execute a certain number of jobs on
a specific time interval(for example 10 per 100ms). Instead of using a
channel to send individual jobs to the pool, just set an atomic number
and run the worker pool until that number reaches 0 (all jobs are
executed).